Give option to upload whole folders
