+++ src/pages/images/index.ts
@@ -1,7 +1,6 @@
 export { Application } from './Application/Application';
 export { DeleteButton } from './DeleteButton/DeleteButton';
 export { DeleteImageDialog } from './DeleteImageDialog/DeleteImageDialog';
-export { FileInput } from './FileInput/FileInput';
 export { Gallery } from './Gallery/Gallery';
 export {
   GalleryCustomDragLayer
